Untested DHT11

+++ app1/shared/src/data.rs
@@ -2,8 +2,8 @@ use wincode::{SchemaRead, SchemaWrite};
 
 #[derive(Default, Clone, SchemaWrite, SchemaRead)]
 pub struct MeteoData {
-    pub temperature: f32,
-    pub humidity: f32,
+    pub temperature: u8,
+    pub humidity: u8,
     pub pressure: f32,
     pub sunlight: f32,
     pub precipitation: f32,
